feat: added jina ai embeddings
