+++ .example.env
@@ -1,2 +1,3 @@
 TIDB_URL=mysql://*******************
 GITHUB_TOKEN=ghp_*******************
+JINA_API_KEY=jina_*****************
