+++ requirements.txt
@@ -9,5 +9,6 @@ streamlit
 llama-index
 llama-index-vector-stores-tidbvector
 llama-index-readers-github
-llama-index-embeddings-huggingface
 llama-index-llms-ollama
+jina
+llama-index-embeddings-jinaai
