+++ src/rag.py
@@ -3,6 +3,9 @@ from dotenv import load_dotenv
 from llama_index.core import StorageContext, VectorStoreIndex
 from llama_index.readers.github import GithubRepositoryReader, GithubClient
 from llama_index.vector_stores.tidbvector import TiDBVectorStore
+from llama_index.embeddings.jinaai import JinaEmbedding
+from llama_index.core import Settings
+from llama_index.llms.ollama import Ollama
 
 # Load environment variables from .env file
 load_dotenv()
@@ -17,6 +20,16 @@ github_client = GithubClient(github_token=github_token, verbose=False)
 tidb_connection_url = os.getenv("TIDB_URL")
 if not tidb_connection_url:
     raise ValueError("TiDB connection URL not found in environment variables.")
+
+# Initialize Jina Embedding Model
+jinaai_api_key = os.getenv("JINA_API_KEY")
+Settings.embed_model = JinaEmbedding(
+    api_key=jinaai_api_key,
+    model="jina-embeddings-v2-base-en"
+)
+
+# initialize llm
+Settings.llm = Ollama(model="llama3", request_timeout=360.0)
 
 async def create_index(owner, repo):
     try:
@@ -53,3 +66,10 @@ async def create_index(owner, repo):
         print(f"An error occurred: {e}")
         return None
 
+
+# https://docs.pingcap.com/tidbcloud/vector-search-integrate-with-llamaindex
+# TODO: research metadata filters and possible use them if they make results better
+def response (index, query):
+    query_engine = index.as_query_engine()
+    response = query_engine.query(query)
+    return response
